dermascan/requirements.txt
django
djangorestframework
Pillow   # for image uploads
